requirements.txt
numpy==2.4.2
opencv-python-headless==4.13.0.92
albumentations==2.0.8
segmentation-models-pytorch==0.5.0
mlflow==3.10.0
pyyaml==6.0.3
tqdm==4.67.3
matplotlib==3.10.8
